requirements.txt
matplotlib==3.1.0
numpy==1.19.5
tensorflow==2.12.0
tensorflow_gpu==2.6.0
pandas==1.1.5
Pillow==9.5.0
scikit-fuzzy==0.4.2
opencv_python==4.6.0.66
scipy==1.4.1
